Fixed security issue regarding passwords
New to 1.3.2, all commands entered by players are echoed in plain text
to the console, which is set up to save to a log file by default.  This
includes the /login and /chpasswd commands, and any passwords supplied
with them.  Added a new listener to intercept these commands before they
reach the console to prevent passwords from being logged in plain text.